CNAM-308 refactor packages
