+++ src/main/scala/fr/polytechnique/cmap/cnam/study/fall/fractures/PrivateAmbulatoryFractures.scala
